Fix implicit type conversion

+++ cf-agent/files_editline.c
@@ -593,7 +593,7 @@ static PromiseResult VerifyLineInsertions(EvalContext *ctx, const Promise *pp, E
     if (preserve_block)
     {
     // promise to insert duplicates on first pass only
-    snprintf(lockname, CF_BUFSIZE - 1, "insertline-%s-%s-%lu", pp->promiser, edcontext->filename, pp->org_pp);
+    snprintf(lockname, CF_BUFSIZE - 1, "insertline-%s-%s-%lu", pp->promiser, edcontext->filename, (long unsigned int) pp->org_pp);
     }
     else
     {
